backend/requirements.txt
fastapi
psycopg2-binary
pydantic
uvicorn